+++ test/mocks/mock.mocks.dart
@@ -2,14 +2,10 @@
 // in twitter_api_v2/test/mocks/mock.dart.
 // Do not manually edit this file.
 
-// Dart imports:
 import 'dart:async' as _i4;
 
-// Package imports:
 import 'package:http/http.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
-
-// Project imports:
 import 'package:twitter_api_v2/src/client/client_context.dart' as _i3;
 import 'package:twitter_api_v2/src/client/user_context.dart' as _i5;
 
@@ -56,11 +52,19 @@ class MockClientContext extends _i1.Mock implements _i3.ClientContext {
           as _i4.Future<_i2.Response>);
   @override
   _i4.Future<_i2.Response> delete(_i5.UserContext? userContext, Uri? uri,
+          {Duration? timeout = const Duration(seconds: 10)}) =>
+      (super.noSuchMethod(
+          Invocation.method(#delete, [userContext, uri], {#timeout: timeout}),
+          returnValue:
+              Future<_i2.Response>.value(_FakeResponse_0())) as _i4
+          .Future<_i2.Response>);
+  @override
+  _i4.Future<_i2.Response> put(_i5.UserContext? userContext, Uri? uri,
           {Map<String, String>? headers = const {},
           dynamic body,
           Duration? timeout = const Duration(seconds: 10)}) =>
       (super.noSuchMethod(
-              Invocation.method(#delete, [userContext, uri],
+              Invocation.method(#put, [userContext, uri],
                   {#headers: headers, #body: body, #timeout: timeout}),
               returnValue: Future<_i2.Response>.value(_FakeResponse_0()))
           as _i4.Future<_i2.Response>);